chore: bundle worker in main script
