+++ package.json
@@ -52,7 +52,8 @@
     "typescript": "3.2.1",
     "webpack": "4.26.1",
     "webpack-cli": "3.1.2",
-    "webpack-dev-server": "^3.1.10"
+    "webpack-dev-server": "^3.1.10",
+    "worker-loader": "^2.0.0"
   },
   "dependencies": {
     "canvas": "^2.2.0",
