+++ webpack.config.js
@@ -10,7 +10,6 @@ module.exports = {
   mode: 'development',
   entry: {
     'imgstry.browser': `${PATH.src}/browser/imgstry/index.js`,
-    'imgstry.worker': `${PATH.src}/browser/worker/index.js`,
     'imgstry.spline': `${PATH.src}/browser/spline/index.js`,
     'imgstry.pixel': `${PATH.src}/pixel/index.js`,
     'imgstry.kernel': `${PATH.src}/kernel/index.js`,
@@ -23,6 +22,20 @@ module.exports = {
     globalObject: `(typeof self !== 'undefined' ? self : this)`,
   },
   devtool: 'source-map',
+  module: {
+    rules: [
+      {
+        test: /\.worker\.js$/,
+        use: {
+          loader: 'worker-loader',
+          options: {
+            inline: true,
+            fallback: false,
+          }
+        }
+      }
+    ]
+  },
   resolve: {
     extensions: ['.js']
   },
